fix(payments): coerce webhook created_at to int before DateTimeImmutable(@) (F22.7)
`WebhookEvent::fromArray()` had `$timestamp = $payload['created_at'] ?? 0`
which only handled the missing-key case. A non-numeric `created_at`
in the JSON payload (attacker-controlled, malformed vendor data,
ISO-8601 string from a misconfigured client) flowed through unchanged
and reached `new DateTimeImmutable('@' . $timestamp)`. On PHP 8.3+
this raises `DateMalformedStringException`, which the upstream
`ProcessWebhookHandler` catches as `JsonException | ValueError` only
— so the date exception escaped to a 500 response instead of the
intended 400 / event-rejection path.

Fix: cast to int before concatenation. PHP semantics map any
non-numeric value to 0 (Unix epoch — recognisable as malformed in
the audit trail without crashing the request) while preserving
numeric strings like Stripe-style `created_at: "1700000000"`.

Add three regression tests:
- non-numeric created_at -> epoch (no exception)
- missing created_at -> epoch (preserves prior behaviour)
- numeric string created_at -> correct timestamp

Closes audit finding F22.7.

+++ extensions/payments/src/Domain/WebhookEvent.php
@@ -41,8 +41,18 @@ final readonly class WebhookEvent
         /** @var string $type */
         $type = $payload['type'] ?? '';
 
-        /** @var int $timestamp */
-        $timestamp = $payload['created_at'] ?? 0;
+        // F22.7: cast to int before concatenation. The previous
+        // `?? 0` short-circuit only protected the missing-key case;
+        // a non-numeric `created_at` ("garbage", "1700-01-01", or any
+        // attacker-controlled JSON value) flowed through as a string
+        // and reached `new DateTimeImmutable('@<string>')`, which on
+        // PHP 8.3+ raises `DateMalformedStringException` and surfaces
+        // as 500 because the upstream handler only catches
+        // JsonException | ValueError. The (int) cast normalises any
+        // non-numeric value to 0 (Unix epoch — recognisable as
+        // malformed) and keeps the constructor signature in the
+        // safe `int|numeric-string` shape it expects.
+        $timestamp = (int) ($payload['created_at'] ?? 0);
 
         /** @var array<string, mixed> $data */
         $data = $payload['data'] ?? [];
